+++ code/frameworks/nextjs/src/routing/decorator.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import type { Addon_StoryContext } from '@storybook/types';
 import { action } from '@storybook/addon-actions';
+import { AppRouterProvider } from './app-router-provider';
 import { PageRouterProvider } from './page-router-provider';
-import type { AppRouterProvider as TAppRouterProvider } from './app-router-provider';
 import type { RouteParams, NextAppDirectory } from './types';
 
 const defaultRouterParams: RouteParams = {
@@ -16,19 +16,6 @@ export const RouterDecorator = (
 ): React.ReactNode => {
   const nextAppDirectory =
     (parameters.nextjs?.appDirectory as NextAppDirectory | undefined) ?? false;
-
-  const [AppRouterProvider, setAppRouterProvider] = React.useState<
-    typeof TAppRouterProvider | undefined
-  >();
-
-  React.useEffect(() => {
-    if (!nextAppDirectory) {
-      return;
-    }
-    import('./app-router-provider').then((exports) =>
-      setAppRouterProvider(() => exports.AppRouterProvider)
-    );
-  }, [nextAppDirectory]);
 
   if (nextAppDirectory) {
     if (!AppRouterProvider) {