+++ code/lib/cli/src/automigrate/fixes/remove-global-client-apis.test.ts
@@ -1,17 +1,16 @@
-/// <reference types="@types/jest" />;
-
 /* eslint-disable no-underscore-dangle */
+import { describe, it, expect, vi } from 'vitest';
+
 import path from 'path';
+import * as fsExtra from 'fs-extra';
 import type { JsPackageManager } from '../../js-package-manager';
 import { RemovedAPIs, removedGlobalClientAPIs as migration } from './remove-global-client-apis';
 
-// eslint-disable-next-line global-require, jest/no-mocks-import
-jest.mock('fs-extra', () => require('../../../../../__mocks__/fs-extra'));
+vi.mock('fs-extra', async () => import('../../../../../__mocks__/fs-extra'));
 
 const check = async ({ contents, previewConfigPath }: any) => {
   if (contents) {
-    // eslint-disable-next-line global-require
-    require('fs-extra').__setMockFiles({
+    vi.mocked<typeof import('../../../../../__mocks__/fs-extra')>(fsExtra as any).__setMockFiles({
       [path.join('.storybook', 'preview.js')]: contents,
     });
   }