+++ code/lib/cli/src/doctor/getMismatchingVersionsWarning.ts
@@ -20,6 +20,10 @@ export function getMismatchingVersionsWarnings(
   installationMetadata?: InstallationMetadata,
   allDependencies?: Record<string, string>
 ): string | undefined {
+  if (!installationMetadata) {
+    return undefined;
+  }
+
   const messages: string[] = [];
   try {
     const frameworkPackageName = Object.keys(installationMetadata?.dependencies || []).find(
@@ -27,7 +31,9 @@ export function getMismatchingVersionsWarnings(
         return Object.keys(frameworkPackages).includes(packageName);
       }
     );
-    const cliVersion = getPrimaryVersion('@storybook/cli', installationMetadata);
+    const cliVersion =
+      getPrimaryVersion('@storybook/cli', installationMetadata) ||
+      getPrimaryVersion('storybook', installationMetadata);
     const frameworkVersion = getPrimaryVersion(frameworkPackageName, installationMetadata);
 
     if (!cliVersion || !frameworkVersion || semver.eq(cliVersion, frameworkVersion)) {
@@ -68,14 +74,19 @@ export function getMismatchingVersionsWarnings(
     );
 
     if (filteredDependencies.length > 0) {
+      const packageJsonSuffix = '(in your package.json)';
       messages.push(
         `Based on your lockfile, these dependencies should be upgraded:`,
         filteredDependencies
           .map(
             ([name, dep]) =>
               `${chalk.hex('#ff9800')(name)}: ${dep[0].version} ${
-                allDependencies?.[name] ? '(in your package.json)' : ''
+                allDependencies?.[name] ? packageJsonSuffix : ''
               }`
+          )
+          .sort(
+            (a, b) =>
+              (b.includes(packageJsonSuffix) ? 1 : 0) - (a.includes(packageJsonSuffix) ? 1 : 0)
           )
           .join('\n')
       );
@@ -85,7 +96,7 @@ export function getMismatchingVersionsWarnings(
       `You can run ${chalk.cyan(
         'npx storybook@latest upgrade'
       )} to upgrade all of your Storybook packages to the latest version.
-      
+
       Alternatively you can try manually changing the versions to match in your package.json. We also recommend regenerating your lockfile, or running the following command to possibly deduplicate your Storybook package versions: ${chalk.cyan(
         installationMetadata?.dedupeCommand
       )}`