+++ code/lib/cli/src/generators/baseGenerator.ts
@@ -282,15 +282,6 @@ export async function baseGenerator(
     Object.keys({ ...packageJson.dependencies, ...packageJson.devDependencies })
   );
 
-  if (!installedDependencies.has('react')) {
-    // we add these here because they are required by addon-essentials > addon-docs
-    addonPackages.push('react');
-  }
-  if (!installedDependencies.has('react-dom')) {
-    // we add these here because they are required by addon-essentials > addon-docs
-    addonPackages.push('react-dom');
-  }
-
   // TODO: We need to start supporting this at some point
   if (type === 'renderer') {
     throw new Error(
@@ -337,7 +328,6 @@ export async function baseGenerator(
   if (builder !== CoreBuilder.Vite && !skipBabel) {
     const frameworksThatNeedBabelConfig = [
       '@storybook/react-webpack5',
-      '@storybook/vue-webpack5',
       '@storybook/vue3-webpack5',
       '@storybook/html-webpack5',
       '@storybook/web-components-webpack5',