+++ code/lib/cli/src/js-package-manager/JsPackageManagerFactory.test.ts
@@ -1,3 +1,4 @@
+import { describe, beforeEach, it, expect, vi } from 'vitest';
 import { sync as spawnSync } from 'cross-spawn';
 import { sync as findUpSync } from 'find-up';
 import path from 'path';
@@ -7,26 +8,32 @@ import { PNPMProxy } from './PNPMProxy';
 import { Yarn1Proxy } from './Yarn1Proxy';
 import { Yarn2Proxy } from './Yarn2Proxy';
 
-jest.mock('cross-spawn');
-const spawnSyncMock = spawnSync as jest.Mock;
+vi.mock('cross-spawn');
+const spawnSyncMock = vi.mocked(spawnSync);
 
-jest.mock('find-up');
-const findUpSyncMock = findUpSync as unknown as jest.Mock;
+vi.mock('find-up');
+const findUpSyncMock = vi.mocked(findUpSync);
 
-describe('JsPackageManagerFactory', () => {
+describe('CLASS: JsPackageManagerFactory', () => {
   beforeEach(() => {
     findUpSyncMock.mockReturnValue(undefined);
+    delete process.env.npm_config_user_agent;
   });
 
-  describe('getPackageManager', () => {
-    describe('return an NPM proxy', () => {
-      it('when `force` option is `npm`', () => {
+  describe('METHOD: getPackageManager', () => {
+    describe('NPM proxy', () => {
+      it('FORCE: it should return a NPM proxy when `force` option is `npm`', () => {
         expect(JsPackageManagerFactory.getPackageManager({ force: 'npm' })).toBeInstanceOf(
           NPMProxy
         );
       });
 
-      it('when all package managers are ok, but only a `package-lock.json` file', () => {
+      it('USER AGENT: it should infer npm from the user agent', () => {
+        process.env.npm_config_user_agent = 'npm/7.24.0';
+        expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(NPMProxy);
+      });
+
+      it('ALL EXIST: when all package managers are ok, but only a `package-lock.json` file is found', () => {
         spawnSyncMock.mockImplementation((command) => {
           // Yarn is ok
           if (command === 'yarn') {
@@ -52,7 +59,7 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
 
         // There is only a package-lock.json
@@ -62,14 +69,19 @@ describe('JsPackageManagerFactory', () => {
       });
     });
 
-    describe('return a PNPM proxy', () => {
-      it('when `force` option is `pnpm`', () => {
+    describe('PNPM proxy', () => {
+      it('FORCE: it should return a PNPM proxy when `force` option is `pnpm`', () => {
         expect(JsPackageManagerFactory.getPackageManager({ force: 'pnpm' })).toBeInstanceOf(
           PNPMProxy
         );
       });
 
-      it('when all package managers are ok, but only a `pnpm-lock.yaml` file', () => {
+      it('USER AGENT: it should infer pnpm from the user agent', () => {
+        process.env.npm_config_user_agent = 'pnpm/7.4.0';
+        expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(PNPMProxy);
+      });
+
+      it('ALL EXIST: when all package managers are ok, but only a `pnpm-lock.yaml` file is found', () => {
         spawnSyncMock.mockImplementation((command) => {
           // Yarn is ok
           if (command === 'yarn') {
@@ -95,7 +107,7 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any as any;
         });
 
         // There is only a pnpm-lock.yaml
@@ -104,9 +116,11 @@ describe('JsPackageManagerFactory', () => {
         expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(PNPMProxy);
       });
 
-      it('when a pnpm-lock.yaml file is closer than a yarn.lock', () => {
+      it('PNPM LOCK IF CLOSER: when a pnpm-lock.yaml file is closer than a yarn.lock', async () => {
         // Allow find-up to work as normal, we'll set the cwd to our fixture package
-        findUpSyncMock.mockImplementation(jest.requireActual('find-up').sync);
+        findUpSyncMock.mockImplementation(
+          (await vi.importActual<typeof import('find-up')>('find-up')).sync
+        );
 
         spawnSyncMock.mockImplementation((command) => {
           // Yarn is ok
@@ -133,18 +147,23 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
         const fixture = path.join(__dirname, 'fixtures', 'pnpm-workspace', 'package');
         expect(JsPackageManagerFactory.getPackageManager({}, fixture)).toBeInstanceOf(PNPMProxy);
       });
     });
 
-    describe('return a Yarn 1 proxy', () => {
-      it('when `force` option is `yarn1`', () => {
+    describe('Yarn 1 proxy', () => {
+      it('FORCE: it should return a Yarn1 proxy when `force` option is `yarn1`', () => {
         expect(JsPackageManagerFactory.getPackageManager({ force: 'yarn1' })).toBeInstanceOf(
           Yarn1Proxy
         );
+      });
+
+      it('USER AGENT: it should infer yarn1 from the user agent', () => {
+        process.env.npm_config_user_agent = 'yarn/1.22.11';
+        expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(Yarn1Proxy);
       });
 
       it('when Yarn command is ok, Yarn version is <2, NPM is ko, PNPM is ko', () => {
@@ -171,7 +190,7 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
 
         // there is no lockfile
@@ -206,7 +225,7 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
 
         // There is a yarn.lock
@@ -215,9 +234,11 @@ describe('JsPackageManagerFactory', () => {
         expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(Yarn1Proxy);
       });
 
-      it('when multiple lockfiles are in a project, prefers yarn', () => {
+      it('when multiple lockfiles are in a project, prefers yarn', async () => {
         // Allow find-up to work as normal, we'll set the cwd to our fixture package
-        findUpSyncMock.mockImplementation(jest.requireActual('find-up').sync);
+        findUpSyncMock.mockImplementation(
+          (await vi.importActual<typeof import('find-up')>('find-up')).sync
+        );
 
         spawnSyncMock.mockImplementation((command) => {
           // Yarn is ok
@@ -244,21 +265,26 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
         const fixture = path.join(__dirname, 'fixtures', 'multiple-lockfiles');
         expect(JsPackageManagerFactory.getPackageManager({}, fixture)).toBeInstanceOf(Yarn1Proxy);
       });
     });
 
-    describe('return a Yarn 2 proxy', () => {
-      it('when `force` option is `yarn2`', () => {
+    describe('Yarn 2 proxy', () => {
+      it('FORCE: it should return a Yarn2 proxy when `force` option is `yarn2`', () => {
         expect(JsPackageManagerFactory.getPackageManager({ force: 'yarn2' })).toBeInstanceOf(
           Yarn2Proxy
         );
       });
 
-      it('when Yarn command is ok, Yarn version is >=2, NPM is ko, PNPM is ko', () => {
+      it('USER AGENT: it should infer yarn2 from the user agent', () => {
+        process.env.npm_config_user_agent = 'yarn/2.2.10';
+        expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(Yarn2Proxy);
+      });
+
+      it('ONLY YARN 2: when Yarn command is ok, Yarn version is >=2, NPM is ko, PNPM is ko', () => {
         spawnSyncMock.mockImplementation((command) => {
           // Yarn is ok
           if (command === 'yarn') {
@@ -282,7 +308,7 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
 
         expect(JsPackageManagerFactory.getPackageManager()).toBeInstanceOf(Yarn2Proxy);
@@ -314,7 +340,7 @@ describe('JsPackageManagerFactory', () => {
           // Unknown package manager is ko
           return {
             status: 1,
-          };
+          } as any;
         });
 
         // There is a yarn.lock
@@ -325,7 +351,7 @@ describe('JsPackageManagerFactory', () => {
     });
 
     it('throws an error if Yarn, NPM, and PNPM are not found', () => {
-      spawnSyncMock.mockReturnValue({ status: 1 });
+      spawnSyncMock.mockReturnValue({ status: 1 } as any);
       expect(() => JsPackageManagerFactory.getPackageManager()).toThrow();
     });
   });