+++ code/lib/core-events/src/index.test.ts
@@ -1,3 +1,4 @@
+import { describe, it, expect } from 'vitest';
 import * as EventsPackageExport from './index';
 import EventsDefaultExport, { CHANNEL_CREATED } from './index';
 
@@ -6,6 +7,8 @@ describe('Core Events', () => {
     expect(EventsPackageExport.CHANNEL_CREATED).toBe('channelCreated');
   });
   it('Should export all values in the default export', () => {
+    // this is intentional, for testing purposes
+    // eslint-disable-next-line import/no-named-as-default-member
     expect(EventsDefaultExport.CHANNEL_CREATED).toBe('channelCreated');
   });
   it('Should export values as named exports', () => {