+++ code/lib/manager-api/src/modules/shortcuts.ts
@@ -15,7 +15,7 @@ export const isMacLike = () =>
 export const controlOrMetaKey = () => (isMacLike() ? 'meta' : 'control');
 
 export function keys<O>(o: O) {
-  return Object.keys(o) as (keyof O)[];
+  return Object.keys(o!) as (keyof O)[];
 }
 
 export interface SubState {
@@ -216,7 +216,7 @@ export const init: ModuleFn = ({ store, fullAPI, provider }) => {
       const shortcuts = api.getShortcutKeys();
       const actions = keys(shortcuts);
       const matchedFeature = actions.find((feature: API_Action) =>
-        shortcutMatchesShortcut(shortcut, shortcuts[feature])
+        shortcutMatchesShortcut(shortcut!, shortcuts[feature])
       );
       if (matchedFeature) {
         api.handleShortcutFeature(matchedFeature, event);
@@ -275,7 +275,7 @@ export const init: ModuleFn = ({ store, fullAPI, provider }) => {
           if (element) {
             try {
               // should be like a channel message and all that, but yolo for now
-              element.contentWindow.focus();
+              element.contentWindow!.focus();
             } catch (e) {
               //
             }
@@ -392,7 +392,7 @@ export const init: ModuleFn = ({ store, fullAPI, provider }) => {
     });
 
     // Also listen to keydown events sent over the channel
-    provider.channel.on(PREVIEW_KEYDOWN, (data: { event: KeyboardEventLike }) => {
+    provider.channel?.on(PREVIEW_KEYDOWN, (data: { event: KeyboardEventLike }) => {
       api.handleKeydownEvent(data.event);
     });
   };