+++ code/lib/preview-api/src/modules/store/autoTitle.test.ts
@@ -1,5 +1,5 @@
+import { describe, it, expect } from 'vitest';
 import { normalizeStoriesEntry } from '@storybook/core-common';
-import { expect } from '@jest/globals';
 
 import { userOrAutoTitleFromSpecifier as userOrAuto } from './autoTitle';