+++ code/lib/types/src/modules/core-common.ts
@@ -157,6 +157,7 @@ export interface LoadOptions {
   packageJson: PackageJson;
   outputDir?: string;
   configDir?: string;
+  cacheKey?: string;
   ignorePreview?: boolean;
   extendServer?: (server: Server) => void;
 }
@@ -170,6 +171,7 @@ export interface CLIOptions {
   enableCrashReports?: boolean;
   host?: string;
   initialPath?: string;
+  exactPort?: boolean;
   /**
    * @deprecated Use 'staticDirs' Storybook Configuration option instead
    */
@@ -395,6 +397,11 @@ export interface StorybookConfigRaw {
      * This will make sure that your story renders the same no matter if docgen is enabled or not.
      */
     disallowImplicitActionsInRenderV8?: boolean;
+
+    /**
+     * Enable asynchronous component rendering in NextJS framework
+     */
+    experimentalNextRSC?: boolean;
   };
 
   build?: TestBuildConfig;