+++ scripts/release/__tests__/write-changelog.test.ts
@@ -1,19 +1,34 @@
+/* eslint-disable jest/no-mocks-import */
 /* eslint-disable global-require */
 /* eslint-disable no-underscore-dangle */
 import path from 'path';
 import dedent from 'ts-dedent';
+import { vi, expect, describe, it, beforeEach } from 'vitest';
+import * as fsExtraImp from 'fs-extra';
 import { run as writeChangelog } from '../write-changelog';
-import * as changesUtils from '../utils/get-changes';
+import * as changesUtils_ from '../utils/get-changes';
 
-// eslint-disable-next-line jest/no-mocks-import
-jest.mock('fs-extra', () => require('../../../code/__mocks__/fs-extra'));
-const fsExtra = require('fs-extra');
+import type * as MockedFSToExtra from '../../../code/__mocks__/fs-extra';
 
-const getChangesMock = jest.spyOn(changesUtils, 'getChanges');
+vi.mock('fs-extra', async () => import('../../../code/__mocks__/fs-extra'));
+vi.mock('../utils/get-changes');
 
-jest.spyOn(console, 'log').mockImplementation(() => {});
-jest.spyOn(console, 'warn').mockImplementation(() => {});
-jest.spyOn(console, 'error').mockImplementation(() => {});
+const changesUtils = vi.mocked(changesUtils_);
+
+const fsExtra = fsExtraImp as unknown as typeof MockedFSToExtra;
+
+beforeEach(() => {
+  vi.restoreAllMocks();
+
+  vi.spyOn(console, 'log').mockImplementation(() => {});
+  vi.spyOn(console, 'warn').mockImplementation(() => {});
+  vi.spyOn(console, 'error').mockImplementation(() => {});
+
+  fsExtra.__setMockFiles({
+    [STABLE_CHANGELOG_PATH]: EXISTING_STABLE_CHANGELOG,
+    [PRERELEASE_CHANGELOG_PATH]: EXISTING_PRERELEASE_CHANGELOG,
+  });
+});
 
 const STABLE_CHANGELOG_PATH = path.join(__dirname, '..', '..', '..', 'CHANGELOG.md');
 const PRERELEASE_CHANGELOG_PATH = path.join(__dirname, '..', '..', '..', 'CHANGELOG.prerelease.md');
@@ -28,10 +43,6 @@ const LATEST_VERSION_PATH = path.join(
 );
 const NEXT_VERSION_PATH = path.join(__dirname, '..', '..', '..', 'docs', 'versions', 'next.json');
 
-beforeEach(() => {
-  jest.clearAllMocks();
-});
-
 const EXISTING_STABLE_CHANGELOG = dedent`## 7.0.0
 
 - Core: Some change`;
@@ -40,14 +51,9 @@ const EXISTING_PRERELEASE_CHANGELOG = dedent`## 7.1.0-alpha.20
 
 - CLI: Super fast now`;
 
-fsExtra.__setMockFiles({
-  [STABLE_CHANGELOG_PATH]: EXISTING_STABLE_CHANGELOG,
-  [PRERELEASE_CHANGELOG_PATH]: EXISTING_PRERELEASE_CHANGELOG,
-});
-
 describe('Write changelog', () => {
   it('should write to stable changelogs and version files in docs', async () => {
-    getChangesMock.mockResolvedValue({
+    changesUtils.getChanges.mockResolvedValue({
       changes: [],
       changelogText: `## 7.0.1
 
@@ -83,7 +89,7 @@ describe('Write changelog', () => {
   });
 
   it('should escape double quotes for json files', async () => {
-    getChangesMock.mockResolvedValue({
+    changesUtils.getChanges.mockResolvedValue({
       changes: [],
       changelogText: `## 7.0.1
 
@@ -122,7 +128,7 @@ describe('Write changelog', () => {
   });
 
   it('should write to prerelase changelogs and version files in docs', async () => {
-    getChangesMock.mockResolvedValue({
+    changesUtils.getChanges.mockResolvedValue({
       changes: [],
       changelogText: `## 7.1.0-alpha.21