+++ scripts/tasks/sandbox-parts.ts
@@ -8,10 +8,12 @@ import {
   ensureDir,
   existsSync,
   pathExists,
+  readFileSync,
   readJson,
   writeJson,
 } from 'fs-extra';
 import { join, resolve, sep } from 'path';
+import JSON5 from 'json5';
 import { createRequire } from 'module';
 import slash from 'slash';
 
@@ -37,6 +39,8 @@ import { JsPackageManagerFactory } from '../../code/lib/cli/src/js-package-manag
 import { workspacePath } from '../utils/workspace';
 import { babelParse } from '../../code/lib/csf-tools/src/babelParse';
 import { CODE_DIRECTORY, REPROS_DIRECTORY } from '../utils/constants';
+import type { TemplateKey } from '../../code/lib/cli/src/sandbox-templates';
+import type { JsPackageManager } from '../../code/lib/cli/src/js-package-manager';
 
 const logger = console;
 
@@ -73,7 +77,7 @@ export const create: Task['run'] = async ({ key, template, sandboxDir }, { dryRu
   }
 };
 
-export const install: Task['run'] = async ({ sandboxDir }, { link, dryRun, debug }) => {
+export const install: Task['run'] = async ({ sandboxDir, key }, { link, dryRun, debug }) => {
   const cwd = sandboxDir;
   await installYarn2({ cwd, dryRun, debug });
 
@@ -91,24 +95,22 @@ export const install: Task['run'] = async ({ sandboxDir }, { link, dryRun, debug
     // of any storybook packages as verdaccio is not able to both proxy to npm and publish over
     // the top. In theory this could mask issues where different versions cause problems.
     await addPackageResolutions({ cwd, dryRun, debug });
-    await configureYarn2ForVerdaccio({ cwd, dryRun, debug });
+    await configureYarn2ForVerdaccio({ cwd, dryRun, debug, key });
 
     // Add vite plugin workarounds for frameworks that need it
     // (to support vite 5 without peer dep errors)
-    if (
-      [
-        'bench-react-vite-default-ts',
-        'bench-react-vite-default-ts-nodocs',
-        'bench-react-vite-default-ts-test-build',
-        'internal-ssv6-vite',
-        'react-vite-default-js',
-        'react-vite-default-ts',
-        'svelte-vite-default-js',
-        'svelte-vite-default-ts',
-        'vue3-vite-default-js',
-        'vue3-vite-default-ts',
-      ].includes(sandboxDir.split(sep).at(-1))
-    ) {
+    const sandboxesNeedingWorkarounds: TemplateKey[] = [
+      'bench/react-vite-default-ts',
+      'bench/react-vite-default-ts-nodocs',
+      'bench/react-vite-default-ts-test-build',
+      'react-vite/default-js',
+      'react-vite/default-ts',
+      'svelte-vite/default-js',
+      'svelte-vite/default-ts',
+      'vue3-vite/default-js',
+      'vue3-vite/default-ts',
+    ];
+    if (sandboxesNeedingWorkarounds.includes(key)) {
       await addWorkaroundResolutions({ cwd, dryRun, debug });
     }
 
@@ -168,7 +170,7 @@ export const init: Task['run'] = async (
 
   switch (template.expected.framework) {
     case '@storybook/angular':
-      await prepareAngularSandbox(cwd);
+      await prepareAngularSandbox(cwd, template.name);
       break;
     default:
   }
@@ -379,21 +381,28 @@ export async function addExtraDependencies({
   cwd,
   dryRun,
   debug,
+  extraDeps,
 }: {
   cwd: string;
   dryRun: boolean;
   debug: boolean;
+  extraDeps?: string[];
 }) {
   // web-components doesn't install '@storybook/testing-library' by default
-  const extraDeps = [
+  const extraDevDeps = [
     '@storybook/jest@next',
     '@storybook/testing-library@next',
     '@storybook/test-runner@next',
   ];
-  if (debug) logger.log('🎁 Adding extra deps', extraDeps);
+  if (debug) logger.log('🎁 Adding extra dev deps', extraDevDeps);
+  let packageManager: JsPackageManager;
   if (!dryRun) {
-    const packageManager = JsPackageManagerFactory.getPackageManager({}, cwd);
-    await packageManager.addDependencies({ installAsDevDependencies: true }, extraDeps);
+    packageManager = JsPackageManagerFactory.getPackageManager({}, cwd);
+    await packageManager.addDependencies({ installAsDevDependencies: true }, extraDevDeps);
+  }
+  if (extraDeps) {
+    if (debug) logger.log('🎁 Adding extra deps', extraDeps);
+    await packageManager.addDependencies({ installAsDevDependencies: false }, extraDeps);
   }
 }
 
@@ -585,7 +594,7 @@ export const extendMain: Task['run'] = async ({ template, sandboxDir }, { disabl
  * In a second step a docs:json script is placed into the package.json to generate the
  * Compodoc documentation.json, which respects symlinks
  * */
-async function prepareAngularSandbox(cwd: string) {
+async function prepareAngularSandbox(cwd: string, templateName: string) {
   const angularJson = await readJson(join(cwd, 'angular.json'));
 
   Object.keys(angularJson.projects).forEach((projectName: string) => {
@@ -607,4 +616,32 @@ async function prepareAngularSandbox(cwd: string) {
   };
 
   await writeJson(packageJsonPath, packageJson, { spaces: 2 });
+
+  // Set tsConfig compilerOptions
+
+  const tsConfigPath = join(cwd, '.storybook', 'tsconfig.json');
+  const tsConfigContent = readFileSync(tsConfigPath, { encoding: 'utf-8' });
+  // This does not preserve comments, but that shouldn't be an issue for sandboxes
+  const tsConfigJson = JSON5.parse(tsConfigContent);
+
+  tsConfigJson.compilerOptions.noImplicitOverride = false;
+  tsConfigJson.compilerOptions.noPropertyAccessFromIndexSignature = false;
+  tsConfigJson.compilerOptions.jsx = 'react';
+  tsConfigJson.compilerOptions.skipLibCheck = true;
+  tsConfigJson.compilerOptions.noImplicitAny = false;
+  tsConfigJson.compilerOptions.strict = false;
+  tsConfigJson.include = [
+    ...tsConfigJson.include,
+    '../template-stories/**/*.stories.ts',
+    // This is necessary since template stories depend on globalThis.components, which Typescript can't look up automatically
+    '../src/stories/**/*',
+  ];
+
+  if (templateName === 'Angular CLI (Version 15)') {
+    tsConfigJson.compilerOptions.paths = {
+      '@angular-devkit/*': ['node_modules/@angular-devkit/*'],
+    };
+  }
+
+  await writeJson(tsConfigPath, tsConfigJson, { spaces: 2 });
 }